requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
python-dotenv
aioboto3

langchain
duckduckgo-search
ddgs
langchain-community
langchain-core
langchain_openai
bs4
yfinance